# sycl: Add reorder to Q6_K mmvq implementation