This PR implements quants reordering for mmvq for Q6_K quantization following the work done for Q4_0 and Q4_K.
These changes give good results on BMG and do not detriment performance on other GPUs.

### Performance impact

All numbers taken with `GGML_SYCL_DISABLE_OPT=0` . 

#### Battlemage B580

| model                    |        size |  params | backend | ngl |   sm | mmap |  test |   this PR         t/s |  master(26b79b6cb)      t/s |
| ------------------------ | ----------: | ------: | ------- | --: | ---: | ---: | ----: | --------------: | ---------------: |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | none |    0 | pp512 | 7421.88 ± 40.31 | 7303.25 ± 190.33 |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | none |    0 | tg128 |   134.93 ± 4.28 |    132.17 ± 7.13 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | none |    0 | pp512 | 7508.23 ± 12.32 |  7543.68 ± 52.94 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | none |    0 | tg128 |   124.74 ± 2.57 |    117.75 ± 2.95 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | none |    0 | pp512 |  2164.47 ± 3.86 |   2156.43 ± 4.42 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | none |    0 | tg128 |    65.13 ± 0.55 |     65.65 ± 0.38 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | none |    0 | pp512 |  2206.63 ± 4.49 |   2202.15 ± 5.80 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | none |    0 | tg128 |    55.04 ± 0.19 |     52.57 ± 0.23 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | none |    0 | pp512 | 5722.94 ± 33.50 |  5688.02 ± 29.39 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | none |    0 | tg128 |    93.32 ± 2.52 |     88.34 ± 2.42 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | none |    0 | pp512 |  3043.55 ± 8.11 |   3047.39 ± 6.17 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | none |    0 | tg128 |    95.46 ± 2.23 |     95.79 ± 2.39 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | none |    0 | pp512 |  3140.52 ± 6.44 |   3154.81 ± 5.22 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | none |    0 | tg128 |    72.71 ± 0.89 |     69.35 ± 0.16 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | none |    0 | pp512 |  1472.53 ± 4.66 |   1468.04 ± 0.47 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | none |    0 | tg128 |    23.48 ± 0.03 |     20.30 ± 0.05 |

#### Lunar Lake

| model                    |        size |  params | backend | ngl |  test |    this PR         t/s |        master(26b79b6cb)      t/s |
| ------------------------ | ----------: | ------: | ------- | --: | ----: | --------------: | --------------: |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | pp512 | 1559.52 ± 38.29 | 1761.94 ± 15.09 |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | tg128 |    58.17 ± 0.59 |    56.42 ± 0.42 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | pp512 | 1775.38 ± 37.60 | 1675.54 ± 32.04 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | tg128 |    42.21 ± 0.19 |    39.81 ± 0.65 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | pp512 |   391.97 ± 4.96 |   433.80 ± 1.21 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | tg128 |    21.61 ± 0.51 |    20.38 ± 0.61 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | pp512 |   492.83 ± 0.55 |   488.92 ± 1.28 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | tg128 |    15.86 ± 0.24 |    14.98 ± 0.16 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | pp512 |  985.69 ± 62.63 |  990.18 ± 10.61 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | tg128 |    29.24 ± 0.15 |    27.48 ± 0.25 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | pp512 |   674.42 ± 0.61 |   665.08 ± 3.18 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | tg128 |    34.34 ± 0.09 |    33.52 ± 0.11 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | pp512 |   743.15 ± 1.86 |   737.78 ± 2.94 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | tg128 |    22.53 ± 0.48 |    22.17 ± 0.30 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | pp512 |  301.50 ± 14.06 |   302.83 ± 1.09 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | tg128 |     7.72 ± 0.05 |     6.06 ± 0.02 |
| llama 8B Q4_K - Medium   |    4.58 GiB |  8.03 B | SYCL    |  99 | pp512 |   411.81 ± 4.68 |   418.31 ± 5.50 |
| llama 8B Q4_K - Medium   |    4.58 GiB |  8.03 B | SYCL    |  99 | tg128 |    14.46 ± 0.06 |    13.40 ± 0.09 |

#### Intel Arc A770

| model                    |        size |  params | backend | ngl |   sm | mmap |  test |            t/s |     master(26b79b6cb)       t/s |
| ------------------------ | ----------: | ------: | ------- | --: | ---: | ---: | ----: | -------------: | -------------: |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | none |    0 | pp512 | 4456.48 ± 8.69 | 4433.16 ± 9.59 |
| qwen2 1.5B Q4_0          | 1013.62 MiB |  1.78 B | SYCL    |  99 | none |    0 | tg128 |   45.60 ± 0.21 |   44.82 ± 0.24 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | none |    0 | pp512 | 4499.64 ± 3.28 | 4460.18 ± 5.36 |
| qwen2 1.5B Q4_K - Medium |    1.04 GiB |  1.78 B | SYCL    |  99 | none |    0 | tg128 |   44.60 ± 0.16 |   43.98 ± 0.14 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | none |    0 | pp512 | 1716.17 ± 1.29 | 1707.70 ± 1.18 |
| llama 7B Q4_0            |    3.57 GiB |  6.74 B | SYCL    |  99 | none |    0 | tg128 |   34.40 ± 0.03 |   34.06 ± 0.02 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | none |    0 | pp512 | 1732.65 ± 2.73 | 1723.82 ± 1.31 |
| llama 7B Q4_K - Medium   |    3.80 GiB |  6.74 B | SYCL    |  99 | none |    0 | tg128 |   32.60 ± 0.27 |   31.71 ± 0.27 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | none |    0 | pp512 | 3661.03 ± 6.91 | 3632.01 ± 6.45 |
| gemma2 2B Q4_K - Medium  |    1.59 GiB |  2.61 B | SYCL    |  99 | none |    0 | tg128 |   39.20 ± 0.31 |   38.16 ± 0.36 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | none |    0 | pp512 | 2464.93 ± 2.58 | 2445.01 ± 2.54 |
| phi3 3B Q4_0             |    2.03 GiB |  3.82 B | SYCL    |  99 | none |    0 | tg128 |   39.98 ± 0.01 |   39.41 ± 0.33 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | none |    0 | pp512 | 2513.13 ± 2.47 | 2492.01 ± 1.80 |
| phi3 3B Q4_K - Medium    |    2.23 GiB |  3.82 B | SYCL    |  99 | none |    0 | tg128 |   34.50 ± 0.30 |   34.14 ± 0.02 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | none |    0 | pp512 | 1031.68 ± 1.04 | 1024.74 ± 1.07 |
| llama 34B Q6_K           |    8.20 GiB | 10.73 B | SYCL    |  99 | none |    0 | tg128 |   17.33 ± 0.11 |   15.12 ± 0.16 |